requirements.txt
pytest
pylint
python-dotenv
psycopg2-binary
pandas
streamlit
matplotlib
seaborn
plotly
redshift_connector
watchdog
openpyxl

